+++ src/std/StringArrayException.java
@@ -1,0 +1,13 @@
+package std;
+
+class StringArrayException extends Exception {
+    private static final String DEFAULT_DESCRIPTION = "Common error";
+
+    public StringArrayException() {
+        super(DEFAULT_DESCRIPTION);
+    }
+
+    public StringArrayException(String string) {
+        super(string);
+    }
+}
