+++ src/std/StringService.java
@@ -1,76 +1,37 @@
 package std;
 
 import java.util.Arrays;
+import java.util.Comparator;
 
-public class StringService {
+class StringService {
 
     public static String[] sort(String[] as, boolean dir) throws StringArrayException {
-        if(as == null || as.length == 0 ) {
+        if (as == null || as.length == 0) {
             throw new StringArrayException("Массив не создан либо пуст.");
         }
-        String[] tempArray = new String[as.length];
-        System.arraycopy(as, 0, tempArray, 0, as.length);
-        for (int i = 0; i < tempArray.length; i++) {
-            tempArray[i] = tempArray[i].toLowerCase();
-        }
-
         if (dir == true) {
-
-            for (int i = 0; i < tempArray.length; i++) {
-                for (int j = i + 1; j < tempArray.length; j++) {
-                    String strTemp = "";
-                    if (tempArray[j].charAt(0) <= tempArray[i].charAt(0)) {
-                        if(tempArray[j].charAt(0) < tempArray[i].charAt(0)) {
-                            strTemp = tempArray[i];
-                            tempArray[i] = tempArray[j];
-                            tempArray[j] = strTemp;
-                            strTemp = as[i];
-                            as[i] = as[j];
-                            as[j] = strTemp;
-                        }
-                        else {
-                            if(tempArray[j].charAt(1) < tempArray[i].charAt(1)) {
-                                strTemp = tempArray[i];
-                                tempArray[i] = tempArray[j];
-                                tempArray[j] = strTemp;
-                                strTemp = as[i];
-                                as[i] = as[j];
-                                as[j] = strTemp;
-                            }
-                        }
-                    }
+            Arrays.sort(as, new Comparator<String>() {
+                public int compare(String s1, String s2) {
+                    return s1.toLowerCase().compareTo(s2.toLowerCase());
                 }
-            }
-        }
-
-        else {
-            for (int i = 0; i < tempArray.length; i++) {
-                for (int j = i + 1; j < tempArray.length; j++) {
-                    String strTemp = "";
-                    if (tempArray[j].charAt(0) >= tempArray[i].charAt(0)) {
-                        if(tempArray[j].charAt(0) > tempArray[i].charAt(0)) {
-                            strTemp = tempArray[i];
-                            tempArray[i] = tempArray[j];
-                            tempArray[j] = strTemp;
-                            strTemp = as[i];
-                            as[i] = as[j];
-                            as[j] = strTemp;
-                        }
-                        else {
-                            if(tempArray[j].charAt(1) > tempArray[i].charAt(1)) {
-                                strTemp = tempArray[i];
-                                tempArray[i] = tempArray[j];
-                                tempArray[j] = strTemp;
-                                strTemp = as[i];
-                                as[i] = as[j];
-                                as[j] = strTemp;
-                            }
-                        }
-                   }
-                }
-            }
-        }
+            });
             return as;
+        }
+        else {
+            Arrays.sort(as, new Comparator<String>() {
+                public int compare(String s1, String s2) {
+                    return s1.toLowerCase().compareTo(s2.toLowerCase());
+                }
+            });
+            String[] tempArray = new String[as.length];
+            System.arraycopy(as,0, tempArray, 0, as.length);
+            int j = tempArray.length-1;
+            for(int i = 0; i < as.length; i++) {
+                as[i] = tempArray[j];
+                j--;
+            }
+            return as;
+        }
     }
 
     public static char[] getCharStat (String str) {
@@ -87,39 +48,4 @@ public class StringService {
 
         return array;
     }
-}
-
-class StringArrayException extends Exception {
-    private static final String DEFAULT_DESCRIPTION = "Common error";
-
-    public StringArrayException() {
-        super(DEFAULT_DESCRIPTION);
-    }
-
-    public StringArrayException(String string) {
-        super(string);
-    }
-
-    public static void main(String[] args) {
-        StringService service = new StringService();
-        char[] array = service.getCharStat("Программирование на java");
-        System.out.println(Arrays.toString(array));
-        String[] arrayString = {"стоит", "Арбуз", "Дорого", "Совсем", "Сочный", "южный", "Ялтинский"};
-        try {
-            arrayString = service.sort(arrayString, true);
-        }
-        catch(StringArrayException e) {
-
-        }
-        System.out.println(Arrays.toString(arrayString));
-
-        try {
-            arrayString = service.sort(arrayString, false);
-        }
-        catch(StringArrayException e) {
-
-        }
-        System.out.println(Arrays.toString(arrayString));
-    }
-
 }
